feat(repository): agrega metodo para buscar correo por el id de un cliente

+++ src/main/java/com/misproyectos/interfaces/CorreoRepInterface.java
@@ -14,5 +14,5 @@ public interface CorreoRepInterface {
 
     CorreoCliente findById(int idCorreo) throws SQLException;
 
-    CorreoCliente findByClientId(int idCliente) throws SQLException;
+    CorreoCliente findByClientId(Long idCliente) throws SQLException;
 }
